feat: add analytics subcomponents
